+++ Tests/UIKitTests/UIViewControllerExtensionsTests.swift
@@ -57,12 +57,20 @@ final class UIViewControllerExtensionsTests: XCTestCase {
     
     func testShowAlert() {
         let viewController = UIViewController()
+        UIApplication.shared.keyWindow?.rootViewController = viewController
         let title = "test title"
         let message = "test message"
-        let alertController = viewController.showAlert(title: title, message: message, buttonTitles: ["OK", "Cancel"], highlightedButtonIndex: 1, completion: nil)
+        let actionButtons = ["OK", "Cancel"]
+        let preferredButtonIndex = 1
+        let alertController = viewController.showAlert(title: title, message: message, buttonTitles: actionButtons, highlightedButtonIndex: preferredButtonIndex, completion: nil)
         XCTAssertEqual(alertController.preferredStyle, .alert)
         XCTAssertEqual(alertController.title, title)
         XCTAssertEqual(alertController.message, message)
+        //check whether the buttons are added in the same order
+        for i in 0..<alertController.actions.count {
+            XCTAssertEqual(alertController.actions[i].title, actionButtons[i])
+        }
+        XCTAssertEqual(alertController.preferredAction, alertController.actions[preferredButtonIndex])
     }
 }
 #endif